Explicit transformation to integer type for nuclei id

+++ scripts/src/utils/general.py
@@ -19,13 +19,13 @@ class hidden_prints:
 def get_dict_from_yaml(pathToYAML):
     with open(str(pathToYAML), 'r') as f:
         yamlContent = f.read()
-    
+
     return yaml.safe_load(yamlContent)
 
 def lnA_stats(A, E, bins):
     lnA = np.log(A)
     E = np.asarray(E)
-    
+
     container = []
     l = len(bins) - 1
     for i in range(l):
@@ -33,19 +33,17 @@ def lnA_stats(A, E, bins):
         lnAi = lnA[inBin]
         element = [lnAi.mean(), lnAi.var()] if (lnAi.size > 0) else [np.nan, np.nan]
         container.append(element)
-    
+
     return np.array(container)
 
 def event_counter(data, bins):
-    
     energies = 18. + np.log10(data['E'])
-    A = np.array([massNumber(id) for id in data['ID'].astype(int)])
+    A = np.array([massNumber(int(id)) for id in data['ID']])
     counts = np.histogram(energies[A >= 1], bins = bins)[0]
-    
+
     return counts, energies, A
 
 def events_from_files(fileNames, bins = pao.ebins):
-    
     lnAContainer = []
     counts = np.zeros(len(bins)-1)
     for fName in fileNames:
@@ -68,16 +66,16 @@ def events_from_files(fileNames, bins = pao.ebins):
 
 def print_args(args):
     argDict = vars(args)
-    
+
     title = "Argument values ({} items)".format(len(argDict))
     print("{0}{1}{0}{2}{0}".format("\n", title.rjust(40), ("="*len(title)).rjust(40)))
-    
+
     for arg, val in argDict.items():
         print("{}:    {}".format(arg.rjust(25), val.resolve() if (isinstance(val, Path) and val.exists()) else val))
     print("\n")
-    
+
 def rescale_from_normal(interval, value):
     a, b = interval
     iSize = b - a
-    
+
     return a + value*iSize
